+++ src/lib/nav/NavUl.svelte
@@ -27,6 +27,6 @@
 [Go to docs](https://svelte-5-ui-lib.codewithshin.com/)
 ## Props
 @props: children: Snippet;
-@props:activeUrl: string = '';
+@props:activeUrl: string = "";
 @props:class: string;
 -->